Implement toast as a service
