+++ app/services/toast.js
@@ -1,0 +1,1 @@
+export { default } from 'ember-toastr/services/toast';
